move aurora types to custom_types
