+++ qdax/core/aurora.py
@@ -13,6 +13,7 @@ from qdax.core.containers.mapelites_repertoire import MapElitesRepertoire
 from qdax.core.containers.unstructured_repertoire import UnstructuredRepertoire
 from qdax.core.emitters.emitter import Emitter, EmitterState
 from qdax.custom_types import (
+    AuroraExtraInfo,
     Descriptor,
     ExtraScores,
     Fitness,
@@ -22,7 +23,6 @@ from qdax.custom_types import (
     Params,
     RNGKey,
 )
-from qdax.tasks.brax.common.descriptor_extractors import AuroraExtraInfo
 
 
 class AURORA:
