+++ qdax/tasks/brax/common/descriptor_extractors.py
@@ -5,7 +5,7 @@ import jax
 import jax.numpy as jnp
 
 from qdax.core.neuroevolution.buffers.buffer import QDTransition
-from qdax.custom_types import Descriptor, Params
+from qdax.custom_types import AuroraExtraInfoNormalization, Descriptor
 
 
 def get_final_xy_position(data: QDTransition, mask: jnp.ndarray) -> Descriptor:
@@ -39,45 +39,6 @@ def get_feet_contact_proportion(data: QDTransition, mask: jnp.ndarray) -> Descri
     descriptors = descriptors / jnp.sum(1.0 - mask, axis=1)
 
     return descriptors
-
-
-class AuroraExtraInfo(flax.struct.PyTreeNode):
-    """
-    Information specific to the AURORA algorithm.
-
-    Args:
-        model_params: the parameters of the dimensionality reduction model
-    """
-
-    model_params: Params
-
-
-class AuroraExtraInfoNormalization(AuroraExtraInfo):
-    """
-    Information specific to the AURORA algorithm. In particular, it contains
-    the normalization parameters for the observations.
-
-    Args:
-        model_params: the parameters of the dimensionality reduction model
-        mean_observations: the mean of observations
-        std_observations: the std of observations
-    """
-
-    mean_observations: jnp.ndarray
-    std_observations: jnp.ndarray
-
-    @classmethod
-    def create(
-        cls,
-        model_params: Params,
-        mean_observations: jnp.ndarray,
-        std_observations: jnp.ndarray,
-    ) -> AuroraExtraInfoNormalization:
-        return cls(
-            model_params=model_params,
-            mean_observations=mean_observations,
-            std_observations=std_observations,
-        )
 
 
 def get_aurora_encoding(
